Moved CommitMetadata into separate file. CHANGELOG_BEGIN CHANGELOG_END
